+++ src/render/postprocessing/ColorUber.ts
@@ -363,7 +363,7 @@ class ColorUberRuntime implements ForwardRenderPipelineFeatureRuntime {
         const parameters = context.pipeline.acquirePassParameters(this.#parameters);
         parameters.configure(source, destination);
         context.pipeline.graph.addPass(this.#pass, parameters);
-        context.resources.replaceColor(destination);
+        context.resources.replaceColor(destination, 'srgb');
     }
 
     destroy(): void {